Updated README with table of contents and project details
Updated the README file to include a new table of contents for improved navigation. This table of contents contains links to different sections of the document, including the challenge and my process. Also added details about the technologies used in this project under 'My process'. Removed the design preview image to keep the document text-based.